+++ CHANGELOG.md
@@ -18,6 +18,11 @@ Added / Changed / Deprecated / Fixed / Removed / Security
 - The new LP solver implementation is now the default alternative.
 - The two classes `LinearSolver.GeneralBuilder` and `LinearSolver.StandardBuilder` are replaced by a common `LinearSolver.Builder`. Consequently the methods `newGeneralBuilder()` and `newStandardBuilder()` are deprecated and replaced by `newBuilder()`.
 
+#### org.ojalgo.scalar
+
+- Re-implemented `Quadruple`'s divide method using primitives only. (Previously delegated to `BigDecimal`.)
+
+
 ## [54.0.0] – 2024-06-06
 
 ### Added